src/common/options: update list of options
This commit adds the following options to
src/common/options/rgw.yaml.in:
  1. rgw_d4n_host
  2. rgw_d4n_port
  3. rgw_topic_persistency_time_to_live
  4. rgw_topic_persistency_max_retries
  5. rgw_topic_persistency_sleep_duration
These options were added in commit
8d37f60c10fedbf6a5ecae52a8275c12d2b14e48 in
https://github.com/ceph/ceph/pull/48879 but that commit was never
backported to the Reef or Quincy release branches. The absence of
these confval branches in those release branches makes the backporting
of PRs#52835 and 52836 fail the docs-build checks. This commit should
obviate those failures.

Signed-off-by: Zac Dover <zac.dover@proton.me>