doc/cephadm: explain different methods of cephadm delivery
Explain that only in Reef and later releases is Ceph distributed as an
executable compiled from source code. This note is to go into Quincy and
only into Quincy, to direct new users of Ceph whom circumstance has
delivered into the hands of Quincy and who might have the wrong idea
that the documentation of Reef and later releases applies to their
release.

Signed-off-by: Zac Dover <zac.dover@proton.me>